+++ aprsd/cmds/dev.py
@@ -8,8 +8,9 @@ import logging
 import click
 from oslo_config import cfg
 
+from aprsd import cli_helper, conf, packets, plugin
 # local imports here
-from aprsd import cli_helper, client, conf, packets, plugin
+from aprsd.client import base
 from aprsd.main import cli
 from aprsd.utils import trace
 
@@ -96,7 +97,7 @@ def test_plugin(
     if CONF.trace_enabled:
         trace.setup_tracing(["method", "api"])
 
-    client.Client()
+    base.APRSClient()
 
     pm = plugin.PluginManager()
     if load_all:
