+++ aprsd/stats/__init__.py
@@ -1,5 +1,5 @@
-from aprsd import client as aprs_client
 from aprsd import plugin
+from aprsd.client import stats as client_stats
 from aprsd.packets import packet_list, seen_list, tracker, watch_list
 from aprsd.plugins import email
 from aprsd.stats import app, collector
@@ -16,5 +16,5 @@ stats_collector.register_producer(tracker.PacketTrack)
 stats_collector.register_producer(plugin.PluginManager)
 stats_collector.register_producer(aprsd.APRSDThreadList)
 stats_collector.register_producer(email.EmailStats)
-stats_collector.register_producer(aprs_client.APRSClientStats)
+stats_collector.register_producer(client_stats.APRSClientStats)
 stats_collector.register_producer(seen_list.SeenList)
